feat: work errorBoundary
