+++ src/assets/Error.module.css
@@ -15,6 +15,17 @@
   font-size: 1.3rem;
 }
 
+.errorBtn {
+  align-self: center;
+  background-color: rgba(255, 0, 38, 0.93);
+  color: #fff;
+  background-image: linear-gradient(92.83deg, #ff7426 0, #f93a13 100%);
+  cursor: pointer;
+  border-style: none;
+  border-radius: 80px;
+  padding: 10px;
+}
+
 .button21 {
   align-items: center;
   background-color: #3eb2fd;
