+++ src/assets/Main.module.css
@@ -11,14 +11,3 @@
 .header {
   display: flex;
 }
-
-.errorBtn {
-  align-self: center;
-  background-color: rgba(255, 0, 38, 0.93);
-  color: #fff;
-  background-image: linear-gradient(92.83deg, #ff7426 0, #f93a13 100%);
-  cursor: pointer;
-  border-style: none;
-  border-radius: 80px;
-  padding: 10px;
-}
