+++ src/assets/index.css
@@ -1,8 +1,8 @@
 body {
   background-image: linear-gradient(
-      120deg,
-      rgba(224, 212, 212, 0.75) 0%,
-      rgba(224, 213, 213, 0.61) 100%
+    120deg,
+    rgba(224, 212, 212, 0.75) 0%,
+    rgba(224, 213, 213, 0.61) 100%
   );
   margin: 0;
 }
@@ -12,7 +12,6 @@ body {
 }
 
 .app {
-  display: flex;
   margin: 2rem;
   width: 100vw;
   min-height: 100vh;
