+++ src/pages/_app.tsx
@@ -3,11 +3,16 @@ import { Provider } from 'react-redux';
 import type { AppProps } from 'next/app';
 import store from '../store/store';
 import '../assets/index.css';
+import ErrorBoundary from '../components/Error/ErrorBoundary';
 
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <Provider store={store}>
-      <Component {...pageProps} />
+      <div className="app">
+        <ErrorBoundary>
+          <Component {...pageProps} />
+        </ErrorBoundary>
+      </div>
     </Provider>
   );
 }
